FIXED: column type

+++ db/migrate/20140329011541_create_clone_incidents.rb
@@ -1,7 +1,7 @@
 class CreateCloneIncidents < ActiveRecord::Migration
   def change
     create_table :clone_incidents do |t|
-      t.integer :similarity
+      t.float :similarity
 
       t.timestamps
     end
